pyproject.toml
[project]
name = "goodeye"
version = "0.1.0"
description = "Public CLI for Goodeye - manage AI workflow skills from the terminal."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [
  { name = "Goodeye Labs" },
]
keywords = ["goodeye", "cli", "ai", "workflows", "mcp"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development",
  "Topic :: Utilities",
]
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "rich>=13",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "pyyaml>=6",
]

[project.scripts]
goodeye = "goodeye_cli.__main__:main"

[project.urls]
Homepage = "https://mcp.goodeyelabs.com"
Repository = "https://github.com/Goodeye-Labs/goodeye-cli"
Issues = "https://github.com/Goodeye-Labs/goodeye-cli/issues"

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "respx>=0.21",
  "ruff>=0.6",
  "pyright>=1.1.380",
  "pre-commit>=3.8",
  "types-pyyaml>=6.0",
]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/goodeye_cli"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "PL", "RUF"]
# PLR0913: too many args is OK for typer commands.
# PLC0415: non-top-level imports for lazy loading / testability.
# PLR2004: magic-number comparisons are common in tests and HTTP-status checks.
# B008: typer / FastAPI-style default-value calls are idiomatic.
ignore = ["PLR0913", "PLC0415", "PLR2004", "B008"]

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
